refactor(domain): move entities to @/domain/entities and update imports
- Moved product.entities.ts, date.entities.ts, currency.entities.ts, and cart.entities.ts to @/domain/entities
- Updated imports to reflect the new file paths
