+++ adapters/cart/database/get-cart-by-user-id.ts
@@ -1,4 +1,4 @@
-import { CartEntity } from "@/domain/cart.entities";
+import { CartEntity } from "@/domain/entities/cart.entities";
 import { getDatabaseCartByUserId } from "@/infra/services/cart/get-database-cart-by-user-id";
 import { handleGetCartDatabase } from "@/adapters/cart/database/handle-get-cart-database";
