+++ adapters/cart/database/map-database-cart-item-to-domain-cart-item.ts
@@ -1,5 +1,5 @@
 import { CurrencyAdapter } from "@/adapters/currency/currency.adapter";
-import { CartItemEntity } from "@/domain/cart.entities";
+import { CartItemEntity } from "@/domain/entities/cart.entities";
 import { CartItemDatabase } from "@/infra/db/types/cart";
 
 export function mapDatabaseCartItemToDomainCartItem(
