+++ adapters/currency/generic/create-currency-entity.ts
@@ -1,4 +1,4 @@
-import { CurrencyEntity } from "@/domain/currency.entities";
+import { CurrencyEntity } from "@/domain/entities/currency.entities";
 
 export function createCurrencyEntity(
   input: string | number
